MT5/service/requirements.txt
numpy
pandas
pymysql
